More of restructuring the codebase
